pyproject.toml
[tool.poetry]
name = "my-google-adk-project"
version = "0.1.0"
description = ""
authors = ["ShreyeshSwargam98 <shreyesh.swargam@bootlabstech.com>"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
google-adk = "^1.14.1"
python-dotenv = "^1.1.1"
google-generativeai = "^0.8.5"
psycopg2 = "^2.9.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
